+++ android/src/main/java/br/com/rsmarques/flutter_branch_sdk/FlutterBranchSdkPlugin.java
@@ -47,9 +47,6 @@ import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugin.common.PluginRegistry.NewIntentListener;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
 
-/**
- * FlutterBranchSdkPlugin
- */
 public class FlutterBranchSdkPlugin implements FlutterPlugin, MethodCallHandler, StreamHandler, NewIntentListener, ActivityAware,
         Application.ActivityLifecycleCallbacks {
     private static final String DEBUG_NAME = "FlutterBranchSDK";
@@ -63,15 +60,23 @@ public class FlutterBranchSdkPlugin implements FlutterPlugin, MethodCallHandler,
     private static final String MESSAGE_CHANNEL = "flutter_branch_sdk/message";
     private static final String EVENT_CHANNEL = "flutter_branch_sdk/event";
     private EventSink eventSink = null;
-    private Map<String, Object> initialData = null;
+    private Map<String, Object> initialParams = null;
+    private BranchError initialError = null;
 
-    /**
-     * Plugin registration.
-     */
+    /**---------------------------------------------------------------------------------------------
+     Plugin registry
+     --------------------------------------------------------------------------------------------**/
 
     public static void registerWith(Registrar registrar) {
+        if (registrar.activity() == null) {
+            // When a background flutter view tries to register the plugin, the registrar has no activity.
+            // We stop the registration process as this plugin is foreground only.
+            return;
+        }
         FlutterBranchSdkPlugin plugin = new FlutterBranchSdkPlugin();
-        plugin.setupChannels(registrar.messenger(), registrar.context());
+        plugin.setupChannels(registrar.messenger(), registrar.activity().getApplicationContext());
+        plugin.setActivity(registrar.activity());
+        registrar.addNewIntentListener(plugin);
     }
 
     @Override
@@ -86,7 +91,6 @@ public class FlutterBranchSdkPlugin implements FlutterPlugin, MethodCallHandler,
 
     private void setupChannels(BinaryMessenger messenger, Context context) {
         this.context = context;
-        this.activity = null;
 
         methodChannel = new MethodChannel(messenger, MESSAGE_CHANNEL);
         eventChannel = new EventChannel(messenger, EVENT_CHANNEL);
@@ -94,7 +98,12 @@ public class FlutterBranchSdkPlugin implements FlutterPlugin, MethodCallHandler,
         methodChannel.setMethodCallHandler(this);
         eventChannel.setStreamHandler(this);
 
-        FlutterBranchSdkInit.init(this.context);
+        FlutterBranchSdkInit.init(context);
+    }
+
+    private void setActivity(Activity activity) {
+        this.activity = activity;
+        activity.getApplication().registerActivityLifecycleCallbacks(this);
     }
 
     private void teardownChannels() {
@@ -103,13 +112,14 @@ public class FlutterBranchSdkPlugin implements FlutterPlugin, MethodCallHandler,
         this.context = null;
     }
 
+    /**---------------------------------------------------------------------------------------------
+     ActivityAware Interface Methods
+     --------------------------------------------------------------------------------------------**/
     @Override
-    public void onAttachedToActivity(ActivityPluginBinding binding) {
-        this.activity = binding.getActivity();
-        this.activityPluginBinding = binding;
-
-        activity.getApplication().registerActivityLifecycleCallbacks(this);
-        binding.addOnNewIntentListener(this);
+    public void onAttachedToActivity(ActivityPluginBinding activityPluginBinding) {
+        this.activityPluginBinding = activityPluginBinding;
+        setActivity(activityPluginBinding.getActivity());
+        activityPluginBinding.addOnNewIntentListener(this);
     }
 
     @Override
@@ -124,25 +134,37 @@ public class FlutterBranchSdkPlugin implements FlutterPlugin, MethodCallHandler,
     }
 
     @Override
-    public void onReattachedToActivityForConfigChanges(ActivityPluginBinding binding) {
-        onAttachedToActivity(binding);
+    public void onReattachedToActivityForConfigChanges(ActivityPluginBinding activityPluginBinding) {
+        onAttachedToActivity(activityPluginBinding);
     }
 
+    /**---------------------------------------------------------------------------------------------
+     StreamHandler Interface Methods
+     --------------------------------------------------------------------------------------------**/
     @Override
     public void onListen(Object o, EventChannel.EventSink eventSink) {
         this.eventSink = eventSink;
-        if (initialData != null) {
-            eventSink.success(initialData);
-            initialData = null;
+        if (initialParams != null) {
+            eventSink.success(initialParams);
+            initialParams = null;
+            initialError = null;
+        } else if (initialError != null) {
+            eventSink.error(String.valueOf(initialError.getErrorCode()), initialError.getMessage(),null);
+            initialParams = null;
+            initialError = null;
         }
     }
 
     @Override
     public void onCancel(Object o) {
         this.eventSink = null;
-        initialData = null;
+        initialError = null;
+        initialParams = null;
     }
 
+    /**---------------------------------------------------------------------------------------------
+     ActivityLifecycleCallbacks Interface Methods
+     --------------------------------------------------------------------------------------------**/
     @Override
     public void onActivityCreated(Activity activity, Bundle bundle) {
     }
@@ -171,6 +193,9 @@ public class FlutterBranchSdkPlugin implements FlutterPlugin, MethodCallHandler,
         }
     }
 
+    /**---------------------------------------------------------------------------------------------
+     NewIntentListener Interface Methods
+     --------------------------------------------------------------------------------------------**/
     @Override
     public boolean onNewIntent(Intent intent) {
         if (this.activity != null) {
@@ -180,35 +205,10 @@ public class FlutterBranchSdkPlugin implements FlutterPlugin, MethodCallHandler,
         }
         return true;
     }
-    //----------------------------------------------------------------------------------------------
 
-    private Branch.BranchReferralInitListener branchReferralInitListener = new
-            Branch.BranchReferralInitListener() {
-                @Override
-                public void onInitFinished(JSONObject params, BranchError error) {
-                    if (error == null) {
-                        Log.d(DEBUG_NAME, "branchReferralInitListener" + params.toString());
-                        if (eventSink == null) {
-                            try {
-                                initialData = paramsToMap(params);
-                            } catch (JSONException e) {
-                                e.printStackTrace();
-                            }
-                            return;
-                        }
-                        try {
-                            eventSink.success(paramsToMap(params));
-                        } catch (JSONException e) {
-                            e.printStackTrace();
-                        }
-                    } else {
-                        Log.d(DEBUG_NAME, "branchReferralInitListener - error: " + error.toString());
-                    }
-                }
-            };
-
-    //----------------------------------------------------------------------------------------------
-
+    /**---------------------------------------------------------------------------------------------
+     MethodCallHandler Interface Methods
+     --------------------------------------------------------------------------------------------**/
     @Override
     public void onMethodCall(@NonNull MethodCall call, @NonNull Result result) {
         switch (call.method) {
@@ -253,10 +253,39 @@ public class FlutterBranchSdkPlugin implements FlutterPlugin, MethodCallHandler,
         }
     }
 
-    //----------------------------------------------------------------------------------------------
+    /**---------------------------------------------------------------------------------------------
+     Branch SDK Call Methods
+     --------------------------------------------------------------------------------------------**/
+    private Branch.BranchReferralInitListener branchReferralInitListener = new
+            Branch.BranchReferralInitListener() {
+                @Override
+                public void onInitFinished(JSONObject params, BranchError error) {
+                    if (error == null) {
+                        Log.d(DEBUG_NAME, "BranchReferralInitListener - params: " + params.toString());
+                        try {
+                            initialParams = paramsToMap(params);
+                        } catch (JSONException e) {
+                            Log.d(DEBUG_NAME, "BranchReferralInitListener - error to Map: " + e.getLocalizedMessage());
+                            return;
+                        }
+                        if (eventSink != null) {
+                            eventSink.success(initialParams);
+                            initialParams = null;
+                        }
+                    } else {
+                        Log.d(DEBUG_NAME, "BranchReferralInitListener - error: " + error.toString());
+                        if (eventSink != null) {
+                            eventSink.error(String.valueOf(error.getErrorCode()), error.getMessage(),null);
+                            initialError = null;
+                        } else {
+                            initialError = error;
+                        }
+                    }
+                }
+            };
 
     private void validateSDKIntegration() {
-        IntegrationValidator.validate(activity);
+        IntegrationValidator.validate(context);
     }
 
     private void getShortUrl(MethodCall call, final Result result) {
@@ -434,6 +463,9 @@ public class FlutterBranchSdkPlugin implements FlutterPlugin, MethodCallHandler,
         Branch.getInstance().disableTracking(value);
     }
 
+    /**---------------------------------------------------------------------------------------------
+     Object Conversion Functions
+     --------------------------------------------------------------------------------------------**/
     private BranchUniversalObject convertToBUO(HashMap<String, Object> argsMap) {
 
         BranchUniversalObject buo = new BranchUniversalObject();