+++ example/lib/main.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_branch_sdk/flutter_branch_sdk.dart';
 
 void main() => runApp(MyApp());
@@ -21,6 +22,7 @@ class _MyAppState extends State<MyApp> {
   var scaffoldKey = new GlobalKey<ScaffoldState>();
   StreamSubscription<Map> streamSubscription;
   StreamController<String> controllerData = StreamController<String>();
+  StreamController<String> controllerInitSession = StreamController<String>();
   StreamController<String> controllerUrl = StreamController<String>();
 
   @override
@@ -33,7 +35,6 @@ class _MyAppState extends State<MyApp> {
 
   void listenDynamicLinks() async {
     streamSubscription = FlutterBranchSdk.initSession().listen((data) {
-      print('initSession');
       print('listenDynamicLinks - DeepLink Data: $data');
       controllerData.sink.add((data.toString()));
       if (data.containsKey('+clicked_branch_link') &&
@@ -51,7 +52,11 @@ class _MyAppState extends State<MyApp> {
             duration: 10);
       }
     }, onError: (error) {
-      print(error);
+      PlatformException platformException = error as PlatformException;
+      print(
+          'InitSession error: ${platformException.code} - ${platformException.message}');
+      controllerInitSession.add(
+          'InitSession error: ${platformException.code} - ${platformException.message}');
     });
   }
 
@@ -159,6 +164,29 @@ class _MyAppState extends State<MyApp> {
         body: ListView(
           padding: EdgeInsets.all(10),
           children: <Widget>[
+            StreamBuilder<String>(
+              stream: controllerInitSession.stream,
+              initialData: '',
+              builder: (context, snapshot) {
+                if (snapshot.hasData && snapshot.data.isNotEmpty) {
+                  return Column(
+                    children: <Widget>[
+                      Center(
+                          child: Text(
+                        snapshot.data,
+                        textAlign: TextAlign.center,
+                        style: TextStyle(
+                            fontSize: 16,
+                            fontWeight: FontWeight.bold,
+                            color: Colors.red),
+                      ))
+                    ],
+                  );
+                } else {
+                  return Container();
+                }
+              },
+            ),
             RaisedButton(
               child: Text('Validate SDK Integration'),
               onPressed: () {
@@ -403,6 +431,7 @@ class _MyAppState extends State<MyApp> {
     super.dispose();
     controllerData.close();
     controllerUrl.close();
+    controllerInitSession.close();
     streamSubscription.cancel();
   }
 }