+++ CHANGELOG.md
@@ -4,4 +4,7 @@
 * Pubspec.yaml Update
 ## 0.1.1
 * Minor adjustments and fix initial deep link data loss
+## 0.1.2
+* Compatibility with apps built on earlier versions of Flutter 1.12
+* Improved error handling in initSession